requirements.txt
pandas==2.2.3
numpy==2.2.1
streamlit==1.41.1
plotly==5.24.1
sqlalchemy==2.0.36
uuid==1.30
psycopg2-binary==2.9.10
python-dotenv==1.0.1